refactor: style코드 분할하고 router를 설정한다
