+++ package.json
@@ -9,7 +9,7 @@
     "@types/jest": "^27.4.1",
     "@types/styled-components": "^5.1.25",
     "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "styled-normalize": "^8.0.7",
     "typescript": "^4.6.3",
@@ -45,6 +45,8 @@
     "@types/node": "^17.0.23",
     "@types/react": "^18.0.1",
     "@types/react-dom": "^18.0.0",
+    "@types/react-router": "^5.1.18",
+    "@types/react-router-dom": "^5.3.3",
     "@typescript-eslint/eslint-plugin": "^5.18.0",
     "@typescript-eslint/parser": "^5.18.0",
     "eslint": "^8.13.0",
