+++ src/common/header/navigation.tsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+import { RightNav } from './styles';
+
+const Navi = function Navi(): JSX.Element {
+  return (
+    <>
+      <RightNav to="/qna">
+        <span>Q&A</span>
+      </RightNav>
+      <RightNav to="/aboutus">
+        <span>About Us</span>
+      </RightNav>
+    </>
+  );
+};
+
+export default Navi;
