+++ src/pages/aboutus.tsx
@@ -1,0 +1,12 @@
+import Header from 'common/header/header';
+
+const AboutUs = function AboutUs(): JSX.Element {
+  return (
+    <>
+      <Header />
+      <span>This is About us Page</span>
+    </>
+  );
+};
+
+export default AboutUs;
