+++ src/pages/qna.tsx
@@ -1,0 +1,12 @@
+import Header from 'common/header/header';
+
+const Qna = function Qna(): JSX.Element {
+  return (
+    <>
+      <Header />
+      <span>This is qna Page</span>
+    </>
+  );
+};
+
+export default Qna;
